renderer working for walls

+++ map.hpp
@@ -15,7 +15,7 @@ class Map{
     void init(Wall ***, int, int, int, int);
     void twoDRender(SDL_Renderer *, int, int, int);
     void threeDRender(SDL_Renderer *, Player*, int = 0, int = 0, int = 640, int = 480);
-    std::tuple<int, int, Wall> calcWDistance(int, int, double, SDL_Renderer *);
+    std::tuple<double, double, Wall> calcWDistance(int, int, double, SDL_Renderer *);
     void defaultMap(int, int);
   private:
     int height, width, bsize, nfloors;
